feat: implemented post component (#2)

+++ app/api/users/[id]/route.ts
@@ -1,5 +1,7 @@
 import prisma from '../../../../prisma/client';
 import { NextResponse } from 'next/server';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/app/lib/auth';
 
 export async function GET(
   request: Request,
@@ -27,15 +29,50 @@ export async function DELETE(
   { params }: { params: { id: string } },
 ) {
   try {
-    const user = await prisma.user.delete({
+    const session = await getServerSession(authOptions);
+
+    const user = await prisma.user.findUnique({
       where: {
         id: params.id,
       },
     });
 
-    return NextResponse.json({
+    if (!user) {
+      return NextResponse.json({ status: 404, error: 'Resource not found' });
+    }
+
+    if (user.id !== session?.user.id && session?.user.role !== 'admin') {
+      return NextResponse.json({
+        status: 401,
+        error: 'You are not authorized to delete this resource',
+      });
+    }
+
+    const results = await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: params.id,
+        },
+      }),
+      prisma.comment.deleteMany({
+        where: {
+          userId: params.id,
+        },
+      }),
+      prisma.post.deleteMany({
+        where: {
+          userId: params.id,
+        },
+      }),
+      prisma.user.delete({
+        where: {
+          id: params.id,
+        },
+      }),
+    ]);
+
+    return NextResponse.json(results, {
       status: 200,
-      message: 'Resource deleted successfully',
     });
   } catch (error) {
     console.error('Error deleting user:', error);
